Rename `instrumentation` to `stopwatch` (#45196)
... since that's what the code is/does.

Partial work towards https://github.com/flutter/flutter/issues/126009.
