+++ ci/licenses_golden/excluded_files
@@ -53,7 +53,6 @@
 ../../../flutter/flow/flow_run_all_unittests.cc
 ../../../flutter/flow/frame_timings_recorder_unittests.cc
 ../../../flutter/flow/gl_context_switch_unittests.cc
-../../../flutter/flow/instrumentation_unittests.cc
 ../../../flutter/flow/layers/backdrop_filter_layer_unittests.cc
 ../../../flutter/flow/layers/checkerboard_layertree_unittests.cc
 ../../../flutter/flow/layers/clip_path_layer_unittests.cc
@@ -75,6 +74,7 @@
 ../../../flutter/flow/mutators_stack_unittests.cc
 ../../../flutter/flow/raster_cache_unittests.cc
 ../../../flutter/flow/skia_gpu_object_unittests.cc
+../../../flutter/flow/stopwatch_unittests.cc
 ../../../flutter/flow/surface_frame_unittests.cc
 ../../../flutter/flow/testing
 ../../../flutter/flow/texture_unittests.cc
