+++ flow/BUILD.gn
@@ -18,8 +18,6 @@ source_set("flow") {
     "embedded_views.h",
     "frame_timings.cc",
     "frame_timings.h",
-    "instrumentation.cc",
-    "instrumentation.h",
     "layer_snapshot_store.cc",
     "layer_snapshot_store.h",
     "layers/backdrop_filter_layer.cc",
@@ -77,6 +75,8 @@ source_set("flow") {
     "raster_cache_util.cc",
     "raster_cache_util.h",
     "skia_gpu_object.h",
+    "stopwatch.cc",
+    "stopwatch.h",
     "surface.cc",
     "surface.h",
     "surface_frame.cc",
@@ -145,7 +145,6 @@ if (enable_unittests) {
       "flow_test_utils.h",
       "frame_timings_recorder_unittests.cc",
       "gl_context_switch_unittests.cc",
-      "instrumentation_unittests.cc",
       "layers/backdrop_filter_layer_unittests.cc",
       "layers/checkerboard_layertree_unittests.cc",
       "layers/clip_path_layer_unittests.cc",
@@ -167,6 +166,7 @@ if (enable_unittests) {
       "mutators_stack_unittests.cc",
       "raster_cache_unittests.cc",
       "skia_gpu_object_unittests.cc",
+      "stopwatch_unittests.cc",
       "surface_frame_unittests.cc",
       "testing/mock_layer_unittests.cc",
       "testing/mock_texture_unittests.cc",
